third_party/blink/web_tests/external/wpt/html/canvas/offscreen/fill-and-stroke-styles/2d.pattern.paint.repeat.outside-expected.txt

This is a testharness.js-based test.
FAIL OffscreenCanvas test: 2d.pattern.paint.repeat.outside assert_equals: Red channel of the pixel at (1, 1) expected 0 but got 255
Harness: the test ran to completion.
